Crate hammer

NAME

Safe Network - hammer Version: 0.0.4

A simple front end to the Safe Network.

Basic Usage

$ hammer

Optional: language code ISO639-1, defaults to en.

$ hammer el
$ hammer en
$ hammer ko

Options:

-h, --help          display help and exit
-V, --version       output version information and exit

DESCRIPTION

This is a simple front end GUI for the Safe Network and makes calls to the Safe Network Command Line Interface (CLI).

For more about the Safe Network see:

https://safenetwork.tech/

https://safenetforum.org/

https://github.com/maidsafe/

AUTHOR

David P Brown

REPORTING BUGS

Report bugs to mailto://bugs@davidpbrown.co.uk

COPYRIGHT

Copyleft!

Known issues:

Errors from the network come with colour via color_eyre and then when put to a log file, those create clutter with escape sequences.

Expected behaviour:

User input is cleaned with Rust’s ammonia::clean_text but otherwise the request is put to the CLI as the user asks it.

The actions requested will continue in background Safe CLI processes, even when the GUI and terminal have been closed. Downloads and log files, will then continue to receive the outputs of requests from the network via those safe processes.

Modules

check_dir_exists

Create directory and any parents

cli_args

Module cli_args handles options and returns other arguments.

galleries

Safe GUI content

img_dragon

Image of dragon as bytes

img_safe_logo

Image of the safe logo as bytes

img_venn_psf

Image of Venn diagram [Privacy; Security; Freedom] as bytes

iso8601

Format timestamps as iso8601

user_settings

Functions

main

Safe GUI